+++ boto3/resources/base.py
@@ -14,6 +14,36 @@
 import boto3
 
 
+class ResourceMeta(object):
+    """
+    An object containing metadata about a resource.
+    """
+    def __init__(self, service_name, identifiers=None, client=None,
+                 data=None):
+        #: (``string``) The service name, e.g. 's3'
+        self.service_name = service_name
+        #: (``list``) List of identifier names
+        self.identifiers = identifiers or []
+        #: (:py:class:`~botocore.client.BaseClient`) Low-level Botocore client
+        self.client = client
+        #: (``dict``) Loaded resource data attributes
+        self.data = data
+
+    def __eq__(self, other):
+        # Two metas are equal if their components are all equal
+        return self.service_name == other.service_name and \
+               self.identifiers == other.identifiers and \
+               self.client == other.client and \
+               self.data == other.data
+
+    def copy(self):
+        """
+        Create a copy of this metadata object.
+        """
+        return ResourceMeta(self.service_name, identifiers=self.identifiers,
+                            client=self.client, data=self.data)
+
+
 class ServiceResource(object):
     """
     A base class for resources.
@@ -29,11 +59,13 @@ class ServiceResource(object):
     from when the instance was hydrated. For example::
 
         # Get a low-level client from a resource instance
-        client = resource.meta['client']
+        client = resource.meta.client
         response = client.operation(Param='foo')
 
         # Print the resource instance's service short name
-        print(resource.meta['service_name'])
+        print(resource.meta.service_name)
+
+    See :py:class:`ResourceMeta` for more information.
     """
 
     def __init__(self, *args, **kwargs):
@@ -43,14 +75,14 @@ class ServiceResource(object):
 
         # Create a default client if none was passed
         if kwargs.get('client') is not None:
-            self.meta['client'] = kwargs.get('client')
+            self.meta.client = kwargs.get('client')
         else:
-            self.meta['client'] = boto3.client(self.meta['service_name'])
+            self.meta.client = boto3.client(self.meta.service_name)
 
         # Allow setting identifiers as positional arguments in the order
         # in which they were defined in the ResourceJSON.
         for i, value in enumerate(args):
-            setattr(self, self.meta['identifiers'][i], value)
+            setattr(self, self.meta.identifiers[i], value)
 
         # Allow setting identifiers via keyword arguments. Here we need
         # extra logic to ignore other keyword arguments like ``client``.
@@ -58,20 +90,20 @@ class ServiceResource(object):
             if name == 'client':
                 continue
 
-            if name not in self.meta['identifiers']:
+            if name not in self.meta.identifiers:
                 raise ValueError('Unknown keyword argument: {0}'.format(name))
 
             setattr(self, name, value)
 
         # Validate that all identifiers have been set.
-        for identifier in self.meta['identifiers']:
+        for identifier in self.meta.identifiers:
             if getattr(self, identifier) is None:
                 raise ValueError(
                     'Required parameter {0} not set'.format(identifier))
 
     def __repr__(self):
         identifiers = []
-        for identifier in self.meta['identifiers']:
+        for identifier in self.meta.identifiers:
             identifiers.append('{0}={1}'.format(
                 identifier, repr(getattr(self, identifier))))
         return "{0}({1})".format(
@@ -86,7 +118,7 @@ class ServiceResource(object):
 
         # Each of the identifiers should have the same value in both
         # instances, e.g. two buckets need the same name to be equal.
-        for identifier in self.meta['identifiers']:
+        for identifier in self.meta.identifiers:
             if getattr(self, identifier) != getattr(other, identifier):
                 return False